backend/requirements.txt
# SmartLensOCR Backend Requirements
# Python 3.9+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Google Generative AI
google-generativeai==0.5.0

# Image Processing
Pillow==10.1.0

# Environment Variables
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.0

# Database
#sqlite3  # Built-in, but listed for reference

# Development (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.12.0
flake8==6.1.0
